% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{The function to classify the suspended data in the last layer}
\usage{
assert(train_data, test_data, prob = 0.5)
}
\arguments{
\item{train_data}{Training dataset used for model generation}

\item{test_data}{A testing dataset to cross-validate the model}

\item{prob}{Prior proportion of the data classified as 1}
}
\description{
This function classifies the data unclassified after the last iteration using glm()
}
\examples{
dataxx = simulate_normal(n=40, p=0.8)
dataxx1 = data.frame(dataxx[2*(1:400), ])
dataxx2 = data.frame(dataxx[2*(1:400)-1, ])
assert(train_data, test_data, prob = 0.5)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{TriCTree,}
\keyword{assert}
