% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_interval.R
\name{mid_interval}
\alias{mid_interval}
\title{This is a hidden function of the TriCTree package}
\usage{
mid_interval(data1, data2, col_i)
}
\arguments{
\item{data1}{A subset of data with the indicator value of 1}

\item{data2}{A subset of data with the indicator value of 0}

\item{col_i}{Index of the column to be separated}
}
\description{
This function estimates the upper and lower bounds of the selection criteria
}
\examples{
dataxx = simulate_normal(n=40, p=0.8)
data1 = data.frame(dataxx[2 * (1:400), ])
data2 = data.frame(dataxx[2 * (1:400)-1, ])
mid_interval(data1, data2, col_i = 2)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{interval,}
\keyword{middle,}
\keyword{select}
