% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte.R
\name{monte}
\alias{monte}
\title{A monte Carlo method for generating nonnormal variables}
\usage{
monte(nvar, nsub, cormat, skewvec, kurtvec)
}
\arguments{
\item{nvar}{Number of variables}

\item{nsub}{Number of simulated subjects (response vectors)}

\item{cormat}{The desired correlation matrix}

\item{skewvec}{A vector of indicator skewness values}

\item{kurtvec}{A vector of indicator kurtosis values}
}
\description{
This function generates nonnormal variables
}
\examples{
R <- matrix(0.5,20,20)
diag(R) <- 1
skewvec=rep(3.5,20)
kurtvec=rep(20,20)
monte(nvar=20, nsub=800,cormat=R,skewvec=skewvec,kurtvec=kurtvec)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{monte-carlo}
\keyword{nonnormal,}
