% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nonnormal.R
\name{simulate_nonnormal}
\alias{simulate_nonnormal}
\title{A monte Carlo method for generating nonnormal}
\usage{
simulate_nonnormal(p = 0.5, skewvec = rep(3.5, 20), kurtvec = rep(20,
  20))
}
\arguments{
\item{p}{correlation between each two variables}

\item{skewvec}{A vector of indicator skewness values}

\item{kurtvec}{A vector of indicator kurtosis values}
}
\description{
This function generates nonnormal variables as training and testing dataset
}
\examples{
simulate_nonnormal(p=0.5, skewvec=rep(3.5,20), kurtvec=rep(20,20))
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{monte-carlo}
\keyword{nonnormal,}
