% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_normal.R
\name{simulate_normal}
\alias{simulate_normal}
\title{A function for generating variables with normal distribution}
\usage{
simulate_normal(p = 0.8, mean = rep(0, 20), n = 40)
}
\arguments{
\item{p}{Correlation between two variables}

\item{mean}{A vector indicating the mean value of classification 0}

\item{n}{Number of datasets to generate, with 20 observations in each dataset}
}
\description{
This function generates normal distributed variables
}
\examples{
dataxx = simulate_normal(p = 0.8, mean = rep(0, 20), n = 40)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{normal}
\keyword{simulate,}
