% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_lrtct}
\alias{summary_lrtct}
\title{A summary function of the classification tree}
\usage{
summary_lrtct(param)
}
\arguments{
\item{param}{Parameters of classification tree}
}
\description{
This function generates nonnormal variables
}
\examples{
dataxx = simulate_normal(n=40, p=0.8)
dataxx1 = data.frame(dataxx[2*(1:400), ])
param = TriCTree_tripart(X1~., dataxx1)
summary_lrtct(param)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{parameter}
\keyword{summary,}
