% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriCTree_predict.R
\name{TriCTree_predict}
\alias{TriCTree_predict}
\title{A function  used for prediction in the TriCTree package}
\usage{
TriCTree_predict(data, data2, param, prob = 0.5)
}
\arguments{
\item{data}{Train dataset to generate the classification tree}

\item{data2}{Test dataset}

\item{param}{The parameters of the classification tree}

\item{prob}{Prior proportion of the data with indicating value of 1}
}
\description{
This function predicts the classification of test data with application of the generated classification tree
}
\examples{
dataxx = simulate_normal(n=40, p=0.8)
data = data.frame(dataxx[2*(1:400), ])
data2 = data.frame(dataxx[2*(1:400)-1, ])
param = TriCTree_tripart(X1~., dataxx1)
TriCTree_predict(data, data2, param=param, prob = 0.5)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{TriCTree,}
\keyword{predict}
