% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriCTree_tripart.R
\name{TriCTree_tripart}
\alias{TriCTree_tripart}
\title{The main function of the TriCTree package}
\usage{
TriCTree_tripart(formula = X1 ~ ., data, minleft = 20, minsplit = 5)
}
\arguments{
\item{formula}{A formula of format Y ~ X1 + X2 + ..., where Y needs to be binary
values 0 or 1, and X1, X2, ... are predictor names in the data}

\item{data}{A training dataset for estimating the trichotomous classification tree}

\item{minleft}{Threshold for making a split on the remaining data to be classified; default minleft = 20}

\item{minsplit}{Threshold for making a split on the suspended data, default minsplit = 5}
}
\description{
This function estimates the optimal trichotomous classification tree
}
\examples{
dataxx = simulate_normal(n=40, p=0.8)
dataxx1 = data.frame(dataxx[2*(1:400), ])
param = TriCTree_tripart(X1~., dataxx1)
}
\author{
Yan Sun, Yanke Zhu, Tao Liu
}
\keyword{TriCTree,}
\keyword{tripart}
