% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripart.R
\name{tripart}
\alias{tripart}
\title{The main function of the TriCTree package}
\usage{
tripart(formula = X1 ~ ., data, minleft = 20, minsplit = 5)
}
\arguments{
\item{formula}{A formula of format Y ~ X1 + X2 + ..., where Y needs to be binary
values 0 or 1, and X1, X2, ... are predictor names in the data}

\item{data}{A training dataset for estimating the trichotomous classification tree}

\item{minleft}{Threshold for making a split on the left}

\item{maxleft}{Threshold for making a split on the right}
}
\description{
This function estimates the optimal trichotomous classification tree
}
\examples{
tripart()
}
\keyword{TriCTree,}
\keyword{tripart}
